+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 1.0.0
+* HTTPServiceClient invoked with arguments instead of a dictionary
+* Enable and match parameters used by `requests`
+
+    Before
+    * `data`, used for both query string and request body
+    * `verify_ssl`
+
+    Now
+    * `params`: dictionary or bytes to be sent in the query string
+    * `data`: dictionary, bytes, or file-like object to send in the body
+    * `verify`: verify the SSL cert
+
+
 ## 0.2.1
 * Save error response content in HTTPServiceError exception